requirements.txt
customtkinter
yt-dlp
Pillow
pyglet
sounddevice